requirements.txt
requests
beautifulsoup4
feedparser
python-dotenv
google-genai
edge-tts
pydub
audioop-lts
podgen
pytz
